refactor(line-series): extract spline math into LineSplineMath helper
Pull the per-segment bezier control-point computation and the first-point
seed out of CoreLineSeries.GetSpline into a static LineSplineMath helper
under Kernel/Drawing. Behavior is preserved verbatim — including the
existing length-term arithmetic — so this is a no-op as confirmed by the
SnapshotTests suite (143/143 pixel-identical, 639/639 CoreTests).

Sets the stage for RangeLineSeries which will reuse the same math twice
per point (High curve + Low curve) without subclassing CoreLineSeries.
